+++ brands/WBG/index.js
@@ -1,1 +1,4 @@
-require('./transformers/web')();
+const pkg = require('./package.json');
+const BRAND = pkg.name.replace('@westpac/', '' ).toUpperCase();
+
+require('../../helpers/transformers/web')( BRAND );
