+++ brands/WBG/package.json
@@ -22,8 +22,5 @@
 	},
 	"main": "dist/index.js",
 	"dependencies": {},
-	"devDependencies": {
-		"cfonts": "^2.4.5",
-		"color": "^3.1.1"
-	}
+	"devDependencies": {}
 }
