+++ package.json
@@ -62,7 +62,7 @@
 		"babel-jest": "^23.6.0",
 		"babel-loader": "^8.0.4",
 		"bolt": "^0.22.6",
-		"cfonts": "^2.3.0",
+		"cfonts": "^2.4.5",
 		"chalk": "^2.4.2",
 		"cypress": "^3.1.3",
 		"docz": "^0.13.7",
@@ -83,11 +83,11 @@
 	"dependencies": {
 		"@emotion/core": "^10.0.4",
 		"@emotion/weak-memoize": "^0.2.2",
-		"color": "^3.1.1",
+		"color": "^3.1.2",
 		"common-tags": "^1.8.0",
-		"lodash.merge": "^4.6.2",
 		"facepaint": "^1.2.1",
 		"focus-trap-react": "^6.0.0",
+		"lodash.merge": "^4.6.2",
 		"mini-svg-data-uri": "^1.1.3",
 		"prettier": "^1.15.3",
 		"prop-types": "^15.6.2",
